+++ packages/apps/explorer/src/components/Footer_rename/style.css.ts
@@ -1,5 +1,5 @@
 import { atoms, responsiveStyle, style, tokens } from '@kadena/kode-ui/styles';
-import { $$pageWidth } from '../Layout/styles.css';
+import { $$pageWidth } from '../Layout_rename/styles.css';
 
 export const footerClass = style([
   atoms({
