+++ packages/apps/explorer/src/components/Header_rename/Header_rename.tsx
@@ -5,8 +5,8 @@ import { useRouter } from '@/hooks/router';
 import type { FC } from 'react';
 import React, { useMemo } from 'react';
 import { useInView } from 'react-intersection-observer';
-import { layoutWrapperClass } from '../Layout/styles.css';
-import { NavBar } from '../Navbar/Navbar';
+import { layoutWrapperClass } from '../Layout_rename/styles.css';
+import { NavBar } from '../Navbar_rename/Navbar_rename';
 import { SearchBarHeader } from '../Search/SearchbarHeader/SearchbarHeader';
 import { StatisticsStack } from '../StatisticsComponent/StatisticsStack/StatisticsStack';
 import { fixedClass, fixedVisibleClass, headerClass } from './styles.css';
