+++ packages/apps/explorer/src/components/Header_rename/styles.css.ts
@@ -1,6 +1,6 @@
 import { atoms, responsiveStyle, token, tokens } from '@kadena/kode-ui/styles';
 import { style } from '@vanilla-extract/css';
-import { $$pageWidth } from '../Layout/styles.css';
+import { $$pageWidth } from '../Layout_rename/styles.css';
 
 export const headerClass = style([
   atoms({
