+++ packages/apps/explorer/src/components/Layout_rename/Layout_rename.tsx
@@ -1,5 +1,5 @@
-import { Footer } from '@/components/Footer/Footer';
-import { Header } from '@/components/Header/Header';
+import { Footer } from '@/components/Footer_rename/Footer_rename';
+import { Header } from '@/components/Header_rename/Header_rename';
 import { Stack } from '@kadena/kode-ui';
 import classNames from 'classnames';
 import type { FC, ReactNode } from 'react';
