+++ packages/apps/explorer/src/components/Logo_rename/MobileLogo.tsx
