+++ packages/apps/explorer/src/context/networksContext.tsx
@@ -1,4 +1,4 @@
-import { useToast } from '@/components/Toasts/ToastContext/ToastContext';
+import { useToast } from '@/components/Toast_rename/ToastContext/ToastContext';
 import type { INetwork } from '@/constants/network';
 import { networkConstants } from '@/constants/network';
 import { useRouter } from '@/hooks/router';
