+++ packages/apps/explorer/src/hooks/router.tsx
@@ -1,4 +1,7 @@
-import { createHref, removeNetworkFromPath } from '@/components/Routing/utils';
+import {
+  createHref,
+  removeNetworkFromPath,
+} from '@/components/Routing_rename/utils';
 import {
   getNetworks,
   selectedNetworkKey,
