+++ packages/apps/explorer/src/hooks/search/utils/account.ts
@@ -1,5 +1,5 @@
 import { useAccountQuery } from '@/__generated__/sdk';
-import { useToast } from '@/components/Toasts/ToastContext/ToastContext';
+import { useToast } from '@/components/Toast_rename/ToastContext/ToastContext';
 import { useEffect, useState } from 'react';
 import type { IHookReturnValue } from '..';
 import {
