+++ packages/apps/explorer/src/pages/[networkSlug]/block/[hash].tsx
@@ -2,10 +2,10 @@ import type { BlockQuery } from '@/__generated__/sdk';
 import { useBlockQuery } from '@/__generated__/sdk';
 import { BlockTransactions } from '@/components/BlockTransactions/BlockTransactions';
 import { DataRenderComponent } from '@/components/DataRenderComponent/DataRenderComponent';
-import { Layout } from '@/components/Layout/Layout';
+import { Layout } from '@/components/Layout_rename/Layout_rename';
 import { loadingData } from '@/components/LoadingSkeleton/loadingData/loadingDataBlockquery';
 import { ValueLoader } from '@/components/LoadingSkeleton/ValueLoader/ValueLoader';
-import { useToast } from '@/components/Toasts/ToastContext/ToastContext';
+import { useToast } from '@/components/Toast_rename/ToastContext/ToastContext';
 import { useQueryContext } from '@/context/queryContext';
 import { block } from '@/graphql/queries/block.graph';
 import { useRouter } from '@/hooks/router';
